refactor: migrate comment API to custom fetch hooks
