+++ src/entities/comment/api/index.ts
@@ -1,6 +1,3 @@
 export * from './endpoint'
 export * from './queryKey'
-export * from './get-comments'
-export * from './post-comment'
-export * from './delete-comment'
-export * from './update-comment'
+export * from './use-comments'
