--- src/entities/comment/api/update-comment.ts
@@ -1,39 +1,0 @@
-import { customAxios } from '@/src/shared/api'
-import { CommentType } from '../model'
-import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { COMMENT_URL } from './endpoint'
-import { COMMENT_QUERY_KEY } from './queryKey'
-
-export const patchComment = async (
-  id: string,
-  contents: string
-): Promise<CommentType> => {
-  try {
-    const response = await customAxios.patch(COMMENT_URL.detail(id), {
-      contents,
-    })
-    return response.data.results
-  } catch (error) {
-    console.error(error)
-    throw error
-  }
-}
-
-export const useUpdateComment = (course_id: string) => {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: ({
-      comment_id,
-      contents,
-    }: {
-      comment_id: string
-      contents: string
-    }) => patchComment(comment_id, contents),
-    onSuccess: () => {
-      queryClient.refetchQueries({
-        queryKey: COMMENT_QUERY_KEY.all(course_id),
-      })
-    },
-  })
-}
